+++ apps/server/src/infra/socketio/index.ts
@@ -1,3 +1,3 @@
 export { Socket } from './types';
 export { WsValidationPipe } from './ws-validation.pipe';
-export { MongoIoAdapter } from './socket-mongodb-ioadapter';
+export { MongoIoAdapter } from './mongodb-ioadapter';
